part1-database-etl/requirements.txt
pandas==2.1.1
numpy==1.26.0
mysql-connector-python==8.1.0
python-dateutil==2.8.2
